requirements.txt
aiodns==4.0.0
aiohttp_socks==0.11.0
async_timeout==5.0.1
Babel==2.18.0
brotli==1.2.0
brotlicffi==1.2.0.1
click==8.3.2
ConfigParser==7.2.0
cryptography==46.0.7
cython==3.2.4
docutils==0.22.4
email_validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.25.2
gunicorn==25.3.0
h2==4.3.0
HTMLParser==0.0.2
hypothesis==6.151.12
ipython==8.12.3
ipywidgets==8.1.8
jnius==1.1.0
keyring==25.7.0
mask==1.0.0
motor==3.7.1
openai==2.31.0
outcome==1.3.0.post0
protobuf==7.34.1
pyOpenSSL==26.0.0
pytest==9.0.3
sniffio==1.3.1
socksio==1.0.0
Sphinx==9.1.0
thread==2.0.6
toml==0.10.2
urllib3_secure_extra==0.1.0
uvloop==0.22.1
winloop==0.6.1
xmlrpclib==1.0.1
zstandard==0.25.0
python-dotenv==1.0.1